add pai

+++ code/game/objects/items/devices/hologram_projector.dm
@@ -2,11 +2,12 @@
 	name = "hologram projector"
 	desc = "It makes a hologram appear...with magnets or something..."
 	icon_state = "shield0"
+	can_take_pai = TRUE
 
 	var/mob/living/simple_animal/hologram/advanced/projector/holoperson = null
 	var/holo_range = 6
 	var/holo_mode = 0
-	var/obj/effect/overlay/holoray/ray		//The link between the projection and the projector.
+	var/obj/effect/overlay/holoray/ray	//The link between the projection and the projector.
 	var/datum/recruiter/recruiter = null
 	var/polling_ghosts = FALSE
 
@@ -14,28 +15,35 @@
 	if(holoperson)
 		qdel(holoperson)
 		holoperson = null
+	if(ray)
+		qdel(ray)
+		ray = null
 	..()
 
 /mob/living/simple_animal/hologram/advanced/projector
 	var/obj/item/device/hologram_projector/projector = null
 	var/proj_turf = null
+	login_text = "You are a hologram. You can perform a few basic functions, and are unable to leave the vicinity of the projector.\
+	\n<span class='danger'>Do not damage the station. Do not harm crew members without their consent.</span>"	
 
 /mob/living/simple_animal/hologram/advanced/projector/Destroy()
 	if(projector)
+		projector.icon_state = "shield0"
+		if(projector.ray)
+			qdel(projector.ray)
 		projector = null
-		qdel(projector.ray)
-		projector.ray = null
 	..()
 
 /obj/item/device/hologram_projector/proc/clear_holo()
-	set_light(0)			//pad lighting (hologram lighting will be handled automatically since its owner was deleted)
+	set_light(0)
 	if(holoperson)
 		visible_message("<span class='warning'>The image of [holoperson] fades away.</span>")
 		animate(holoperson, alpha = 0, time = 5)
 		spawn(5)
+			holoperson.set_light(0)
 			qdel(ray)
 			ray = null
-			qdel(holoperson)//Get rid of hologram.
+			qdel(holoperson)
 			holoperson = null
 			icon_state = "shield0"
 	return 1
@@ -45,9 +53,11 @@
 		clear_holo()
 	
 /obj/item/device/hologram_projector/attack_self()
-	if(holoperson)
+	if(integratedpai || holoperson)
 		to_chat(usr, "Shutting down hologram...")
 		clear_holo()
+		if(integratedpai)
+			remove_pai()
 	else
 		to_chat(usr, "Generating hologram...")
 		recruit_holoperson()
@@ -88,11 +98,16 @@
 	player.forceMove(T)
 	var/mob/living/simple_animal/hologram/advanced/projector/H = player.transmogrify(/mob/living/simple_animal/hologram/advanced/projector, TRUE)
 	holoperson = H
+	holoperson.set_light(1)
 	holoperson.projector = src
 	holoperson.proj_turf = T
+
 	icon_state = "shield1"
 	ray = new(T)
 	qdel(recruiter)
+
+/obj/item/device/hologram_projector/examine(mob/user, var/size = "")
+	recruit_holoperson()
 
 /mob/living/simple_animal/hologram/advanced/projector/Life()
 	regular_hud_updates()
@@ -116,11 +131,6 @@
 			return 1
 
 	projector.clear_holo() //If not, we want to get rid of the hologram.
-
-/mob/living/simple_animal/hologram/advanced/projector/Login()
-	login_text = "You are a hologram. You can perform a few basic functions, and are unable to leave the vicinity of the projector.\
-		\n<span class='danger'>Do not damage the station. Do not harm crew members without their consent.</span>"
-	..()
 	
 /mob/living/simple_animal/hologram/advanced/projector/Move()
 	..()
@@ -152,3 +162,26 @@
 			animate(projector.ray, transform = turn(M.Scale(1,sqrt(distx*distx+disty*disty)),newangle),time = 1)
 		else
 			projector.ray.transform = turn(M.Scale(1,sqrt(distx*distx+disty*disty)),newangle)
+
+//pAI it uses the pAI framework in objs.dm. Check that code for further information
+/obj/item/device/hologram_projector/install_pai(obj/item/device/paicard/P)
+	..()
+	if(holoperson)
+		clear_holo()
+	if(!P?.pai)
+		return
+	var/turf/T = get_turf(src)
+	P.pai.forceMove(T)
+	var/mob/living/simple_animal/hologram/advanced/projector/H = P.pai.transmogrify(/mob/living/simple_animal/hologram/advanced/projector, TRUE)
+	holoperson = H
+	holoperson.projector = src
+	holoperson.proj_turf = T
+
+	icon_state = "shield1"
+	ray = new(T)
+
+/obj/item/device/hologram_projector/state_controls_pai(obj/item/device/paicard/P)
+	if(P?.pai && integratedpai)
+		holoperson.login_text = "Your supplemental directives have been updated. Your new directives are: \
+			\nPrime Directive : <br>[P.pai.pai_law0] \
+			\nSupplemental Directives: <br>[P.pai.pai_laws]"
